Move FontAwesomeIcon.svelte to widgets directory
