+++ app/frontend/widgets/FontAwesomeIcon.svelte
